requirements.txt
streamlit
tensorflow
Pillow
numpy
streamlit-drawable-canvas
